Revert back SDM data type to int16_t
For some reason in commit 40b488103ff0a2963786f49bc721050861979d0c
SDM data type was changed from int16_t to uint16_t.
Have no idea why, but it should be int16_t

Fix some warnings

+++ shell.c
@@ -267,7 +267,7 @@ int shell_make_argv (const char *cmdline, char ***argv, int *argc)
 
 	while ((s = next (&cmdline, &quot)) != NULL) {
 		if (quot) {
-			if (*argc == cap) {
+			if ((size_t)*argc == cap) {
 				cap *= 2;
 				*argv = reallocarray (*argv, cap + 1, sizeof (char *));
 			}
